add association between student and classroom
