+++ lib/classroom.rb
@@ -1,0 +1,14 @@
+class Classroom
+  attr_accessor :label
+  attr_reader :students
+
+  def initialize(label)
+    @label = label
+    @students = []
+  end
+
+  def add_student(student)
+    student.classroom = self
+    @students.push(student) unless @students.include?(student)
+  end
+end
