+++ lib/person.rb
@@ -53,3 +53,4 @@ class TrimmerDecorator < BaseDecorator
     @nameable.correct_name.slice(0, 10)
   end
 end
+
